% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_gssm.R, R/generics.R
\name{as_gssm.SSModel}
\alias{as_gssm}
\alias{as_gssm.SSModel}
\title{Coerce to gssm Object}
\usage{
\method{as_gssm}{SSModel}(model, kappa = 1e+05, ...)

as_gssm(model, ...)
}
\arguments{
\item{model}{Object of class \code{SSModel}.}

\item{kappa}{For \code{SSModel} object, a prior variance for initial state u
sed to replace exact diffuse elements of the original model.}

\item{...}{Ignored.}
}
\value{
Object of class \code{gssm}.
}
\description{
Converts Gaussian \code{SSModel} object of \code{KFAS} package to \code{gssm}
object.
}

