% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmc.R
\name{asymptotic_var}
\alias{asymptotic_var}
\title{Asymptotic Variance of IS-type Estimators}
\usage{
asymptotic_var(x, w)
}
\arguments{
\item{x}{Vector of samples.}

\item{w}{Vector of weights.}
}
\description{
Estimates the asymptotic variance based on Corollary 1 
of Vihola et al. (2020) from weighted samples from IS-MCMC.
}
