% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.predict_bssm.R
\name{autoplot.predict_bssm}
\alias{autoplot.predict_bssm}
\title{Plot predictions based on bssm package}
\usage{
\method{autoplot}{predict_bssm}(object, plot_mean = TRUE,
  plot_median = TRUE, ynew, fit, obs_colour = "black",
  mean_colour = "red", median_colour = "blue", fit_colour = "red",
  interval_colour = "#000000", alpha_fill = 0.25, ...)
}
\arguments{
\item{object}{Object of class \code{predict_bssm}.}

\item{...}{Ignored.}
}
\description{
Plot predictions based on bssm package
}

