% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsm}
\alias{bsm}
\title{Basic Structural (Time Series) Model}
\usage{
bsm(y, sd_y, sd_level, sd_slope, sd_seasonal, beta, xreg = NULL,
  period = frequency(y), slope = TRUE, seasonal = frequency(y) > 1, a1,
  P1)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{sd_y}{Prior for the standard error of observation equation. 
See \link{\code{priors}} for details.}

\item{sd_level}{Prior for the standard error of the noise in level equation. 
See \link{\code{priors}} for details. If missing, \code{sd_level} is fixed to zero.}

\item{sd_slope}{Prior for the standard error  of the noise in slope equation. 
See \link{\code{priors}} for details. If missing, \code{sd_slope} is fixed to zero.
Ignored if \code{slope = FALSE}.}

\item{sd_seasonal}{Prior for the standard error of the noise in seasonal equation.
See \link{\code{priors}} for details. If missing, \code{sd_seasonal} is fixed to zero.}

\item{beta}{Prior for the regression coefficients.}

\item{xreg}{Matrix containing covariates.}

\item{period}{Length of the seasonal component i.e. the number of
observations per season. Default is \code{frequency(y)}.}

\item{slope}{Should the model contain the slope term. Default is \code{TRUE}.
If \code{FALSE}, local level component is used instead of local trend component.}

\item{seasonal}{Should the model contain seasonal term. Default is
\code{frequency(y) > 1}.}

\item{a1}{Prior means for the initial states (level, slope, seasonals).
Defaults to vector of zeros.}

\item{P1}{Prior covariance for the initial states (level, slope, seasonals).
Default is diagonal matrix with 1000 on the diagonal.}
}
\value{
Object of class \code{bsm}.
}
\description{
Constructs a basic structural model with local level or local trend component
and seasonal component.
}
\examples{

prior <- uniform(0.1 * sd(log10(UKgas)), 0, 1)
model <- bsm(log10(UKgas), sd_y = prior, 
sd_level =  prior, sd_slope =  prior, sd_seasonal =  prior)

mcmc_out <- run_mcmc(model, n_iter = 5000)
summary(mcmc_out$theta)$stat
mcmc_out$theta[which.max(mcmc_out$posterior), ]
sqrt((fit <- StructTS(log10(UKgas), type = "BSM"))$coef)[c(4, 1:3)]

}

