% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_arguments.R
\name{check_y}
\alias{check_y}
\alias{check_period}
\alias{check_distribution}
\alias{check_sd}
\alias{check_xreg}
\alias{check_beta}
\alias{check_mu}
\alias{check_rho}
\alias{check_phi}
\alias{check_u}
\alias{check_prior}
\alias{check_prop}
\alias{check_D}
\alias{check_C}
\alias{create_regression}
\alias{check_Z}
\alias{check_T}
\alias{check_R}
\alias{check_a1}
\alias{check_P1}
\alias{check_H}
\alias{check_integer}
\alias{check_positive_real}
\alias{check_theta}
\alias{check_missingness}
\title{Check Arguments}
\usage{
check_y(x, multivariate = FALSE, distribution = "gaussian")

check_period(x, n)

check_distribution(x, distribution)

check_sd(x, type, add_prefix = TRUE)

check_xreg(x, n)

check_beta(x, k)

check_mu(x)

check_rho(x)

check_phi(x)

check_u(x, y, multivariate = FALSE)

check_prior(x, name)

check_prop(x, name = "target")

check_D(x, p, n)

check_C(x, m, n)

create_regression(beta, xreg, n)

check_Z(x, p, n, multivariate = FALSE)

check_T(x, m, n)

check_R(x, m, n)

check_a1(x, m)

check_P1(x, m)

check_H(x, p, n, multivariate = FALSE)

check_integer(x, name = "particles", positive = TRUE, max = 1e+07)

check_positive_real(x, name)

check_theta(x)

check_missingness(x)
}
\arguments{
\item{multivariate}{Logical, should \code{p} be larger than 1?}

\item{distribution}{Distribution(s) of the responses.}

\item{n}{Integer, number of time points.}

\item{type}{Name to be added to the sd parameter name.}

\item{add_prefix}{Logical, add \code{type} to parameter name.}

\item{k}{Integer, number of predictors.}

\item{y}{The response time series.}

\item{name}{Name of the argument used in printing error messages.}

\item{p}{Integer, number of time series.}

\item{m}{Integer, dimensionality of the state vector.}

\item{beta}{Vector of regression coefficients.}

\item{xreg}{Matrix or vector of predictors.}

\item{positive}{Logical, check for positiveness of \code{x}.}

\item{max}{Maximum value of \code{x}.}
}
\description{
Check Arguments
}
