% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpp_example_models.R
\name{cpp_example_model}
\alias{cpp_example_model}
\title{Example C++ Codes for Non-Linear and SDE Models}
\usage{
cpp_example_model(example, return_code = FALSE)
}
\arguments{
\item{example}{Name of the example model.
Run \code{cpp_example_model("abc")} to get the names of possible models.}

\item{return_code}{If TRUE, will not compile the model but only returns the
corresponding code.}
}
\value{
Returns pointers to the C++ snippets defining the model, or in case
of \code{return_code = TRUE}, returns the example code without compiling.
}
\description{
Example C++ Codes for Non-Linear and SDE Models
}
\examples{
cpp_example_model("sde_poisson_OU", return_code = TRUE)

}
