% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmc.R
\name{expand}
\alias{expand}
\title{Expand the Jump Chain representation}
\usage{
expand(x, variable = "theta", times, states, by_states = TRUE)
}
\arguments{
\item{x}{Output from \code{\link{run_mcmc}}.}

\item{...}{Ignored.}
}
\description{
The MCMC algorithms of \code{bssm} use a jump chain representation where we 
store the accepted values and the number of times we stayed in the current value.
Although this safes bit memory and is especially convinient for IS-corrected 
MCMC, often we want to have the usual sample path. Function \code{expand} 
returns the expanded sample based on the counts.
}

