% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymptotic_var.R
\name{iact}
\alias{iact}
\title{Integrated Autocorrelation Time}
\usage{
iact(x)
}
\arguments{
\item{x}{A vector.}
}
\description{
Estimates the integrated autocorrelation time based on Sokal (1997).
}
\examples{
set.seed(1)
x <- numeric(1e4)
phi <- 0.8
for(t in 2:1e4) x[t] <- phi * x[t-1] + rnorm(1)
# ESS:
length(x) / iact(x)
}
\references{
Sokal A. (1997) Monte Carlo Methods in Statistical Mechanics: Foundations
and New Algorithms.
In: DeWitt-Morette C., Cartier P., Folacci A. (eds) Functional Integration.
NATO ASI Series (Series B: Physics), vol 361. Springer, Boston, MA.
https://doi.org/10.1007/978-1-4899-0319-8_6
}
