% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ng_bsm.R
\name{importance_sample}
\alias{importance_sample}
\alias{importance_sample.ng_bsm}
\title{Importance Sampling from non-Gaussian State Space Model}
\usage{
importance_sample(object, nsim, seed, ...)

\method{importance_sample}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\description{
Returns \code{nsim} samples from the approximating Gaussian model with corresponding (scaled) importance weights.
}

