% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R, R/gssm.R, R/ng_bsm.R, R/ngssm.R, R/svm.R
\name{kfilter.bsm}
\alias{kfilter}
\alias{kfilter.bsm}
\alias{kfilter.gssm}
\alias{kfilter.ng_bsm}
\alias{kfilter.ngssm}
\alias{kfilter.svm}
\title{Kalman Filtering}
\usage{
\method{kfilter}{bsm}(object, ...)

kfilter(object, ...)

\method{kfilter}{gssm}(object, ...)

\method{kfilter}{ng_bsm}(object, ...)

\method{kfilter}{ngssm}(object, ...)

\method{kfilter}{svm}(object, ...)
}
\arguments{
\item{object}{Model object}

\item{...}{Ignored.}
}
\value{
List containing the log-likelihood (approximate in non-Gaussian case),
one-step-ahead predictions \code{at} and filtered
estimates \code{att} of states, and the corresponding variances \code{Pt} and
 \code{Ptt}.
}
\description{
Function \code{kfilter} runs the Kalman filter for the given model
(and it's parameters, including regression coefficients), and returns the filtered estimates and
one-step-ahead predictions of the states \eqn{\alpha_t} given the
model parameters and data up to time \eqn{t}.
}
\details{
For non-Gaussian models, the Kalman filtering is based on the approximate Gaussian model.
}
\seealso{
\code{\link{particle_filter}}
}

