% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/gssm.R, R/ng_bsm.R, R/ngssm.R
\name{logLik.bstsm}
\alias{logLik.bstsm}
\alias{logLik.gssm}
\alias{logLik.ng_bstsm}
\alias{logLik.ngssm}
\title{Log-likelihood of Exponential Family State Space Model}
\usage{
\method{logLik}{bstsm}(object, ...)

\method{logLik}{gssm}(object, ...)

\method{logLik}{ng_bstsm}(object, ...)

\method{logLik}{ngssm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gssm} or \code{bstsm}.}

\item{...}{Ignored.}
}
\description{
Computes the log-likelihood of exponential family state space model.
}

