% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLik.ssm_gsv}
\alias{logLik.ssm_gsv}
\title{Log-likelihood of a General SV Model}
\usage{
\method{logLik}{ssm_gsv}(
  object,
  particles,
  method = "psi",
  max_iter = 100,
  conv_tol = 1e-08,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)
}
\arguments{
\item{object}{Model model.}

\item{particles}{Number of samples for particle filter or importance sampling. If 0, 
approximate log-likelihood based on the Gaussian approximation is returned.}

\item{method}{Sampling method, default is psi-auxiliary filter (\code{"psi"}), 
other choice us \code{"bsf"}, bootstrap particle filter.}

\item{max_iter}{Maximum number of iterations for Gaussian approximation algorithm.}

\item{conv_tol}{Tolerance parameter for the approximation algorithm.}

\item{seed}{Seed for the random number generator.}

\item{...}{Ignored.}
}
\description{
Computes the log-likelihood of a a general SV model of \code{bssm} package.
}
