% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_bsm.R
\name{ng_bsm}
\alias{ng_bsm}
\title{Non-Gaussian Basic Structural (Time Series) Model}
\usage{
ng_bsm(y, sd_level, sd_slope, sd_seasonal, sd_noise, distribution, phi = 1,
  beta, xreg = NULL, period = frequency(y), slope = TRUE,
  seasonal = frequency(y) > 1, a1, P1)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{sd_level}{Prior for the standard error of the noise in level equation.
See\link[=uniform]{priors} for details. If missing, \code{sd_level} is fixed to zero.}

\item{sd_slope}{Prior for the standard error  of the noise in slope equation.
See\link[=uniform]{priors} for details. If missing, \code{sd_slope} is fixed to zero.
Ignored if \code{slope = FALSE}.}

\item{sd_seasonal}{Prior for the standard error of the noise in seasonal equation.
See\link[=uniform]{priors} for details. If missing, \code{sd_seasonal} is fixed to zero.}

\item{sd_noise}{Prior for the standard error of the additional noise term.
See\link[=uniform]{priors} for details. If missing, no additional noise term is used.}

\item{distribution}{distribution of the observation. Possible choices are
\code{"poisson"} and \code{"binomial"}.}

\item{phi}{Additional parameter vector relating to the non-Gaussian distribution.
For Poisson distribution, this corresponds to offset term. For binomial, this
is the number of trials.}

\item{beta}{Prior for the regression coefficients.}

\item{xreg}{Matrix containing covariates.}

\item{period}{Length of the seasonal component i.e. the number of
observations per season. Default is \code{frequency(y)}.}

\item{slope}{Should the model contain the slope term. Default is \code{TRUE}.
If \code{FALSE}, local level component is used instead of local trend component.}

\item{seasonal}{Should the model contain seasonal term. Default is
\code{frequency(y) > 1}.}

\item{a1}{Prior means for the initial states (level, slope, seasonals).
Defaults to vector of zeros.}

\item{P1}{Prior covariance for the initial states (level, slope, seasonals).
Default is diagonal matrix with 1e5 on the diagonal.}
}
\value{
Object of class \code{ng_bsm}.
}
\description{
Constructs a non-Gaussian basic structural model with local level or
local trend component, a seasonal component, and regression component
(or subset of these components).
}
\examples{
model <- ng_bsm(Seatbelts[, "VanKilled"], distribution = "poisson",
  sd_level = halfnormal(0.01, 1),
  sd_seasonal = halfnormal(0.01, 1), slope = FALSE,
  beta = normal(0, 0, 10),
  xreg = Seatbelts[, "law"])
\dontrun{
set.seed(123)
mcmc_out <- run_mcmc(model, n_iter = 5000, nsim = 10)
mcmc_out$acceptance_rate
plot(mcmc_out$theta)
summary(mcmc_out$theta)

library("ggplot2")
ggplot(as.data.frame(mcmc_out$theta[,1:2]), aes(x = sd_level, y = sd_seasonal)) +
  geom_point() + stat_density2d(aes(fill = ..level.., alpha = ..level..),
  geom = "polygon") + scale_fill_continuous(low = "green",high = "blue") +
  guides(alpha = "none")

pred <- predict(model, n_iter = 5000, nsim_states = 10, n_ahead = 36,
  probs = seq(0.05, 0.95, by = 0.05), newdata = matrix(1, 36, 1),
  newphi = rep(1, 36))
autoplot(pred)
}
}

