% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlg_example_models.R
\name{nlg_example_models}
\alias{nlg_example_models}
\title{Example C++ Codes for Non-Linear Models}
\usage{
nlg_example_models(example, return_code = FALSE)
}
\arguments{
\item{example}{Name of the example model. 
Run \code{nlg_example_models("abc")} to get the names of possible models.}

\item{return_code}{If TRUE, will not compile the model but only returns the 
corresponding code.}
}
\value{
Returns pointers to the C++ snippets defining the model, or in case 
of \code{return_code = TRUE}, returns the example code without compiling.
}
\description{
Example C++ Codes for Non-Linear Models
}
