% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_filter.R
\name{particle_filter}
\alias{particle_filter}
\alias{particle_filter.ngssm}
\alias{particle_filter.nlg_ssm}
\alias{particle_filter.svm}
\title{Particle Filtering}
\usage{
particle_filter(object, nsim, ...)

\method{particle_filter}{ngssm}(object, nsim, filter_type = "bsf",
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_filter}{svm}(object, nsim, filter_type = "bsf",
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_filter}{nlg_ssm}(object, nsim, filter_type = "bsf",
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{of class \code{bsm}, \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{...}{Ignored.}

\item{filter_type}{Eiher \code{"bsf"} or \code{"psi"}.}

\item{seed}{Seed for RNG.}
}
\value{
A list containing samples, weights from the last time point, and an
estimate of log-likelihood.
}
\description{
Function \code{particle_filter} performs a bootstrap filtering with stratification
resampling. For non-Gaussian models, psi-auxiliary particle filter based on 
the Gaussian approximating model is also available.
}

