% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R
\name{particle_simulate.bsm}
\alias{particle_simulate}
\alias{particle_simulate.bsm}
\title{Backwards sampling}
\usage{
\method{particle_simulate}{bsm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

particle_simulate(object, nsim, nsim_store, seed, ...)
}
\arguments{
\item{object}{of class \code{bsm}, \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples used in particle filtering.}

\item{nsim_store}{Number of samples to store.}

\item{seed}{Seed for RNG.}

\item{...}{Ignored.}
}
\description{
Function \code{particle_simulate} performs backwards sampling from
the conditional distribution of the states $p(alpha|y, theta)$.
}

