% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particle_smoother.R
\name{particle_smoother}
\alias{particle_smoother}
\alias{particle_smoother.gssm}
\alias{particle_smoother.ngssm}
\title{Particle Smoothing}
\usage{
particle_smoother(object, nsim, ...)

\method{particle_smoother}{gssm}(object, nsim, smoothing_method = "fs",
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_smoother}{ngssm}(object, nsim, smoothing_method = "fs",
  filter_type = "bootstrap", seed = sample(.Machine$integer.max, size = 1),
  ...)
}
\arguments{
\item{object}{Model.}

\item{nsim}{Number of samples.}

\item{...}{Ignored.}

\item{smoothing_method}{Either \code{"fs"} (filter-smoother), or \code{"fbs"} 
(forward-backward smoother).}

\item{seed}{Seed for RNG.}

\item{filter_type}{Either \code{"bootstrap"} for bootstrap filter, or \code{"psi"}
for psi-auxiliary filter. Latter is only applicable for non-Gaussian/non-linear models.}
}
\description{
Function \code{particle_smoother} performs filter-smoother or forward-backward smoother,
using a either bootstrap filtering or psi-auxiliary filter with stratification resampling.
}

