% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssm-package.R
\docType{data}
\name{poisson_series}
\alias{poisson_series}
\title{Simulated Poisson time series data}
\format{A vector of length 100}
\description{
See example for code for reproducing the data.
}
\examples{
# The data is generated as follows:
set.seed(321)
slope <- cumsum(c(0, rnorm(99, sd = 0.01)))
y <- rpois(100, exp(cumsum(slope + c(0, rnorm(99, sd = 0.1)))))
}
\keyword{datasets}
