% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.mcmc_output}
\alias{predict.bsm}
\alias{predict.mcmc_output}
\alias{predict.ng_bsm}
\alias{predict.ngssm}
\title{Predictions for Gaussian State Space Models}
\usage{
\method{predict}{mcmc_output}(object, future_model, type = "response",
  intervals = TRUE, probs = c(0.05, 0.95), return_MCSE = TRUE,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{predict}{bsm}(object, n_iter, newdata = NULL, n_ahead = 1,
  interval = "response", probs = c(0.05, 0.95), method = "quantile",
  return_MCSE = TRUE, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{predict}{ngssm}(object, n_iter, nsim_states, newdata = NULL,
  n_ahead = 1, interval = "mean", probs = c(0.05, 0.95),
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, seed = sample(.Machine$integer.max, size =
  1), newu = NULL, ...)

\method{predict}{ng_bsm}(object, n_iter, nsim_states, newdata = NULL,
  n_ahead = 1, interval = "mean", probs = c(0.05, 0.95),
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, seed = sample(.Machine$integer.max, size =
  1), newu = NULL, ...)
}
\arguments{
\item{object}{Model object.}

\item{probs}{Desired quantiles. Defaults to \code{c(0.05, 0.95)}. Always includes median 0.5.}

\item{return_MCSE}{For method \code{"parametric"}, if \code{TRUE}, the Monte Carlo
standard errors are also returned.}

\item{...}{Ignored.}

\item{newdata}{Matrix containing the covariate values for the future time
points. Defaults to zero matrix of appropriate size.}

\item{n_ahead}{Number of steps ahead at which to predict.}

\item{interval}{Compute predictions on \code{"mean"} ("confidence interval") or
\code{"response"} ("prediction interval"). Defaults to \code{"response"}.}

\item{method}{Either \code{"parametric"} (default) or \code{"quantile"}.
Only used for linear-Gaussian case.}

\item{nsim_states}{Number of samples used in importance sampling.}

\item{newu}{Vector of length \code{n_ahead} defining the future values of \eqn{u}.
Defaults to 1.}

\item{priors}{Priors for the unknown parameters.}
}
\value{
List containing the mean predictions, quantiles and Monte Carlo
standard errors .
}
\description{
Posterior intervals of future observations or their means
(success probabilities in binomial case) for Gaussian models. These are
computed using either the quantile method where the intervals are computed
as empirical quantiles the posterior sample, or parametric method by
Helske (2016).
}

