% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmc.R
\name{print.mcmc_output}
\alias{print.mcmc_output}
\title{Title}
\usage{
\method{print}{mcmc_output}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{run_mcmc}}.}

\item{...}{Ignored.}
}
\description{
Title
}

