% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_bsm.R, R/ngssm.R, R/svm.R
\name{run_mcmc.ng_bsm}
\alias{run_mcmc.ng_bsm}
\alias{run_mcmc.ngssm}
\alias{run_mcmc.svm}
\title{Bayesian Inference of State Space Models using MCMC}
\usage{
\method{run_mcmc}{ng_bsm}(object, n_iter, nsim_states, type = "full",
  method = "PM", simulation_method = "IS", correction_method = "IS2",
  delayed_acceptance = TRUE, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S, end_adaptive_phase = TRUE,
  adaptive_approx = TRUE, n_threads = 1,
  seed = sample(.Machine$integer.max, size = 1), log_space = FALSE, ...)

\method{run_mcmc}{ngssm}(object, n_iter, Z_est, T_est, R_est, priors,
  nsim_states = 1, n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, end_adaptive_phase = TRUE,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{run_mcmc}{svm}(object, n_iter, nsim_states = 1, type = "full",
  method = "PM", simulation_method = "IS", correction_method = "IS2",
  delayed_acceptance = TRUE, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S, end_adaptive_phase = TRUE,
  adaptive_approx = TRUE, n_threads = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{Model object.}

\item{n_iter}{Number of MCMC iterations.#'}

\item{nsim_states}{Number of states samples per MCMC iteration.}

\item{type}{Either \code{"full"} (default), or \code{"summary"}. The
former produces samples of states whereas the latter gives the mean and
variance estimates of the states.}

\item{method}{Use \code{"standard"} MCMC or \code{"delayed acceptance"} approach.}

\item{simulation_method}{If \code{"IS"} non-sequential importance sampling based
on Gaussian approximation is used. If \code{"PF"} particle filtering (bootstrap filter)
is used.}

\item{correction_method}{For IS correction method, which correction method
should be used? Possible choices are \code{"IS1"}, \code{"IS2"} and \code{"PF"}.
See references for details.}

\item{delayed_acceptance}{For pseudo-marginal MCMC, should delayed acceptance based
on the Gaussian approximation be used?}

\item{n_burnin}{Length of the burn-in period which is disregarded from the
results. Defaults to \code{n_iter / 2}.}

\item{n_thin}{Thinning rate. Defaults to 1. Increase for large models in
order to save memory.}

\item{gamma}{Tuning parameter for the adaptation of RAM algorithm. Must be
between 0 and 1 (not checked).}

\item{target_acceptance}{Target acceptance ratio for RAM. Defaults to 0.234.}

\item{S}{Initial value for the lower triangular matrix of RAM
algorithm, so that the covariance matrix of the Gaussian proposal
distribution is \eqn{SS'}.}

\item{seed}{Seed for the random number generator.}

\item{log_space}{Generate proposals for standard deviations in log-space. Default is \code{TRUE}.}

\item{...}{Ignored.}

\item{Z_est, T_est, R_est}{Matrices or arrays with same dimensions as the
corresponding system matrices in \code{object}, where \code{NA} values
identify the unknown parameters for estimation.}

\item{priors}{Priors for the unknown parameters.}

\item{method}{Whether pseudo-marginal MCMC (\code{"PM"}) (default) or
importance sampling type correction (\code{"IS"}) is used.}

\item{nsim_states}{Number of simulations of states per MCMC iteration. Only
used when \code{type = "full"}.}
}
\description{
Adaptive Markov chain Monte Carlo simulation of state space models using
Robust Adaptive Metropolis algorithm by Vihola (2012).
}

