% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{fast_smoother}
\alias{fast_smoother}
\alias{sim_smoother}
\alias{smoother}
\title{Kalman Smoothing}
\usage{
fast_smoother(object, ...)

smoother(object, ...)

sim_smoother(object, nsim, seed, ...)
}
\arguments{
\item{object}{Model object.}

\item{...}{Ignored.}

\item{nsim}{Number of independent samples.}

\item{seed}{Seed for the random number generator.}
}
\value{
Matrix containing the smoothed estimates of states, list
with the smoothed states and the variances, or an array containing the
generated samples.
}
\description{
Methods for Kalman smoothing of the states. Function \code{fast_smoother}
computes only smoothed estimates of the states, and function
\code{smoother} computes also smoothed variances. Function
\code{sim_smoother} performs simulation smoothing i.e. simulates the states
from the conditional distribution \eqn{p(\alpha | y, \theta)}.
}
\details{
For non-Gaussian models, the smoothing is based on the approximate Gaussian model.
}
\examples{
model <- bsm(rep(NA, 50), sd_level = uniform(1,0,5), sd_y = uniform(1,0,5), slope = FALSE)
sim <- sim_smoother(model, 12)
ts.plot(sim[, 1, ])
}

