% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ssm_gsv}
\alias{ssm_gsv}
\title{Stochastic Volatility Model with Stochastic Mean and Variance processes}
\usage{
ssm_gsv(
  y,
  Z_mu,
  T_mu,
  R_mu,
  D_mu,
  C_mu,
  a1_mu,
  P1_mu,
  Z_sv,
  T_sv,
  R_sv,
  D_sv,
  C_sv,
  a1_sv,
  P1_sv,
  init_theta = numeric(0),
  state_names,
  update_fn = default_update_fn,
  prior_fn = default_prior_fn
)
}
\arguments{
\item{y}{Observations as time series (or vector) of length \eqn{n}.}

\item{Z_mu}{System matrix Z of the mean equation. Either a vector of length m,
a m x n matrix, or object which can be coerced to such.}

\item{T_mu}{System matrix T of the mean process. Either a m x m matrix or a
m x m x n array, or object which can be coerced to such.}

\item{R_mu}{Matrix R of the mean process. Either a m x k matrix or a
m x k x n array, or object which can be coerced to such.}

\item{D_mu}{Intercept terms \eqn{d_t} for the mean equation, given as a
scalar or vector of length n.}

\item{C_mu}{Intercept terms \eqn{c_t} for the state equation for the mean process, given as a
m times 1 or m times n matrix.}

\item{a1_mu}{Prior mean for the initial states of the mean process as a vector of length m.}

\item{P1_mu}{Prior covariance matrix for the initial states of the mean process as m x m matrix.}

\item{Z_sv}{System matrix Z of the variance equation. Either a vector of length m,
a m x n matrix, or object which can be coerced to such.}

\item{T_sv}{System matrix T of the variance process. Either a m x m matrix or a
m x m x n array, or object which can be coerced to such.}

\item{R_sv}{Matrix R of the variance process. Either a m x k matrix or a
m x k x n array, or object which can be coerced to such.}

\item{D_sv}{Intercept terms \eqn{d_t} for the variance equation, given as a
scalar or vector of length n.}

\item{C_sv}{Intercept terms \eqn{c_t} for the state equation for the variance process, given as a
m times 1 or m times n matrix.}

\item{a1_sv}{Prior mean for the initial states of the variance process as a vector of length m.}

\item{P1_sv}{Prior covariance matrix for the initial states of the variance process as m x m matrix.}

\item{init_theta}{Initial values for the unknown hyperparameters theta.}

\item{state_names}{List of length two with vectors defining the names for the states 
of the mean process and variance process.}

\item{update_fn}{Function which returns list of updated model 
components given input vector theta. See details.}

\item{prior_fn}{Function which returns log of prior density 
given input vector theta.}
}
\value{
Model of class \code{gsv}.
}
\description{
Stochastic Volatility Model with Stochastic Mean and Variance processes
}
