% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfilter.R
\name{ukf}
\alias{ukf}
\title{Unscented Kalman Filtering}
\usage{
ukf(object, alpha = 1, beta = 0, kappa = 2)
}
\arguments{
\item{object}{Model object}
}
\value{
List containing the log-likelihood,
one-step-ahead predictions \code{at} and filtered
estimates \code{att} of states, and the corresponding variances \code{Pt} and
 \code{Ptt}.
}
\description{
Function \code{ukf} runs the unscented Kalman filter for the given 
non-linear Gaussian model of class \code{nlg_ssm}, 
and returns the filtered estimates and one-step-ahead predictions of the 
states \eqn{\alpha_t} given the data up to time \eqn{t}.
}

