% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.mcmc_output.R
\name{as.data.frame.mcmc_output}
\alias{as.data.frame.mcmc_output}
\title{Convert MCMC Output to data.frame}
\usage{
\method{as.data.frame}{mcmc_output}(
  x,
  row.names,
  optional,
  variable = c("theta", "states"),
  times,
  states,
  expand = TRUE,
  use_times = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class \code{mcmc_output} from \code{\link{run_mcmc}}.}

\item{row.names}{Ignored.}

\item{optional}{Ignored.}

\item{variable}{Return samples of \code{"theta"} (default) or
\code{"states"}?}

\item{times}{Vector of indices. In case of states,
what time points to return? Default is all.}

\item{states}{Vector of indices. In case of states,
what states to return? Default is all.}

\item{expand}{Should the jump-chain be expanded?
Defaults to \code{TRUE}.
For \code{expand = FALSE} and always for IS-MCMC,
the resulting data.frame contains variable weight (= counts * IS-weights).}

\item{use_times}{If \code{TRUE} (default), transforms the values of the time
variable to match the ts attribute of the input to define. If \code{FALSE},
time is based on the indexing starting from 1.}

\item{...}{Ignored.}
}
\description{
Converts the MCMC output of \code{\link{run_mcmc}} to \code{data.frame}.
}
\examples{
data("poisson_series")
model <- bsm_ng(y = poisson_series, 
sd_slope = halfnormal(0.1, 0.1), 
sd_level = halfnormal(0.1, 1),
  distribution = "poisson")
  
out <- run_mcmc(model, iter = 2000, particles = 10)
head(as.data.frame(out, variable = "theta"))
head(as.data.frame(out, variable = "state"))

# don't expand the jump chain:
head(as.data.frame(out, variable = "theta", expand = FALSE))

# IS-weighted version:
out_is <- run_mcmc(model, iter = 2000, particles = 10, 
  mcmc_type  = "is2")
head(as.data.frame(out_is, variable = "theta"))

}
\seealso{
\code{as_draws} which converts the output for
\code{as_draws} object.
}
