% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_filter.R
\name{bootstrap_filter}
\alias{bootstrap_filter}
\alias{bootstrap_filter.ssm_ulg}
\alias{bootstrap_filter.bsm_lg}
\alias{bootstrap_filter.ssm_ung}
\alias{bootstrap_filter.bsm_ng}
\alias{bootstrap_filter.svm}
\alias{bootstrap_filter.ar1_ng}
\alias{bootstrap_filter.nlg_ssm}
\alias{bootstrap_filter.sde_ssm}
\title{Bootstrap Filtering}
\usage{
bootstrap_filter(object, nsim, ...)

\method{bootstrap_filter}{ssm_ulg}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{bsm_lg}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{ssm_ung}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{bsm_ng}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{svm}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{ar1_ng}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{nlg_ssm}(
  object,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{sde_ssm}(
  object,
  nsim,
  L,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)
}
\arguments{
\item{object}{of class \code{bsm_lg}, \code{bsm_ng} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{...}{Ignored.}

\item{seed}{Seed for RNG.}

\item{L}{Integer defining the discretization level for SDE models.}
}
\value{
A list containing samples, weights from the last time point, and an
estimate of log-likelihood.
}
\description{
Function \code{bootstrap_filter} performs a bootstrap filtering with stratification
resampling.
}
