% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_filter.R
\name{bootstrap_filter}
\alias{bootstrap_filter}
\alias{bootstrap_filter.gaussian}
\alias{bootstrap_filter.nongaussian}
\alias{bootstrap_filter.nlg_ssm}
\alias{bootstrap_filter.sde_ssm}
\title{Bootstrap Filtering}
\usage{
bootstrap_filter(model, nsim, ...)

\method{bootstrap_filter}{gaussian}(
  model,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{nongaussian}(
  model,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{nlg_ssm}(
  model,
  nsim,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)

\method{bootstrap_filter}{sde_ssm}(
  model,
  nsim,
  L,
  seed = sample(.Machine$integer.max, size = 1),
  ...
)
}
\arguments{
\item{model}{of class \code{bsm_lg}, \code{bsm_ng} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{...}{Ignored.}

\item{seed}{Seed for RNG.}

\item{L}{Integer defining the discretization level for SDE models.}
}
\value{
A list containing samples, weights from the last time point, and an
estimate of log-likelihood.
}
\description{
Function \code{bootstrap_filter} performs a bootstrap filtering with stratification
resampling.
}
