% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R
\name{bsm}
\alias{bsm}
\title{Basic Structural (Time Series) Model}
\usage{
bsm(y, sd_y, sd_level, sd_slope, sd_seasonal, beta, xreg = NULL,
  period = frequency(y), slope = TRUE, seasonal = frequency(y) > 1, a1,
  P1)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{sd_y}{Standard error of observation equation. Used as an initial
value in MCMC.}

\item{sd_level}{Standard error of level equation. Used as an initial
value in MCMC. If missing, \code{sd_level} is fixed to zero.}

\item{sd_slope}{Standard error of slope equation. Used as an initial
value in MCMC. If missing, \code{sd_slope} is fixed to zero.
Ignored if \code{slope = FALSE}.}

\item{sd_seasonal}{Standard error of seasonal equation. Used as an initial
value in MCMC. If missing, \code{sd_seasonal} is fixed to zero.}

\item{beta}{Regression coefficients. Used as an initial
value in MCMC. Defaults to vector of zeros.}

\item{xreg}{Matrix containing covariates.}

\item{period}{Length of the seasonal component i.e. the number of
observations per season. Default is \code{frequency(y)}.}

\item{slope}{Should the model contain the slope term. Default is \code{TRUE}.
If \code{FALSE}, local level component is used instead of local trend component.}

\item{seasonal}{Should the model contain seasonal term. Default is
\code{frequency(y) > 1}.}

\item{a1}{Prior means for the initial states (level, slope, seasonals).
Defaults to vector of zeros.}

\item{P1}{Prior covariance for the initial states (level, slope, seasonals).
Default is diagonal matrix with 1000 on the diagonal.}
}
\value{
Object of class \code{bsm}.
}
\description{
Constructs a basic structural model with local level or local trend component
and seasonal component.
}
\examples{

init_sd <- 0.1 * sd(log10(UKgas))
model <- bsm(log10(UKgas), sd_y = init_sd, sd_level = init_sd,
  sd_slope = init_sd, sd_seasonal = init_sd)

mcmc_out <- run_mcmc(model, n_iter = 5000)
summary(mcmc_out$theta)$stat
mcmc_out$theta[which.max(mcmc_out$logLik), ]
sqrt((fit <- StructTS(log10(UKgas), type = "BSM"))$coef)

}

