% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssm-package.R
\docType{package}
\name{bssm}
\alias{bssm}
\title{Bayesian Inference of State Space Models}
\description{
This package contains functions for Bayesian inference of basic stochastic volatility model
and exponential family state space models, where the state equation is linear and Gaussian,
and the conditional observation density is either Gaussian, Poisson,
binomial, negative binomial or Gamma density. General non-linear Gaussian models and models 
with continuous SDE dynamics are also supported. For formal definition of the
currently supported models and methods, as well as some theory behind the IS-MCMC and \eqn{\psi}{psi}-APF, 
see the package vignettes and Vihola, Helske, Franks (2020)
}
\references{
Vihola, M, Helske, J, Franks, J. Importance sampling type estimators based on approximate marginal Markov chain Monte Carlo. 
Scand J Statist. 2020; 1– 38. https://doi.org/10.1111/sjos.12492
}
