% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priors.R
\name{uniform_prior}
\alias{uniform_prior}
\alias{bssm_prior_list}
\alias{uniform}
\alias{halfnormal_prior}
\alias{halfnormal}
\alias{normal_prior}
\alias{normal}
\alias{tnormal_prior}
\alias{tnormal}
\alias{gamma_prior}
\alias{gamma}
\title{Prior objects for bssm models}
\usage{
uniform_prior(init, min, max)

uniform(init, min, max)

halfnormal_prior(init, sd)

halfnormal(init, sd)

normal_prior(init, mean, sd)

normal(init, mean, sd)

tnormal_prior(init, mean, sd, min = -Inf, max = Inf)

tnormal(init, mean, sd, min = -Inf, max = Inf)

gamma_prior(init, shape, rate)

gamma(init, shape, rate)
}
\arguments{
\item{init}{Initial value for the parameter, used in initializing the model 
components and as a starting values in MCMC.}

\item{min}{Lower bound of the uniform and truncated normal prior.}

\item{max}{Upper bound of the uniform and truncated normal prior.}

\item{sd}{Standard deviation of the (underlying i.e. non-truncated) 
Normal distribution.}

\item{mean}{Mean of the Normal prior.}

\item{shape}{Shape parameter of the Gamma prior.}

\item{rate}{Rate parameter of the Gamma prior.}
}
\value{
object of class \code{bssm_prior} or \code{bssm_prior_list} in case 
of multiple priors (i.e. multiple regression coefficients).
}
\description{
These simple objects of class \code{bssm_prior} are used to construct a 
prior distributions for the some of the model objects of \code{bssm} 
package. Currently supported priors are uniform 
(\code{uniform()}), half-normal (\code{halfnormal()}), 
normal (\code{normal()}), gamma (\code{gamma}), and 
truncated normal distribution  (\code{tnormal()}). All parameters are 
vectorized so for regression coefficient vector beta you can define prior 
for example as \code{normal(0, 0, c(10, 20))}.
}
\details{
The longer name versions of the prior functions with \code{_prior} ending 
are identical with shorter versions and they are available only to 
avoid clash with R's primitive function \code{gamma} (other long prior names 
are just for consistent naming).
}
\examples{
# create uniform prior on [-1, 1] for one parameter with initial value 0.2:
uniform(0.2, -1, 1)
# two normal priors at once i.e. for coefficients beta:
normal(init = c(0.1, 2), mean = 0, sd = c(1, 2))
}
