% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssm-package.R
\docType{data}
\name{drownings}
\alias{drownings}
\title{Deaths by drowning in Finland in 1969-2014}
\format{
A time series object containing 46 observations and.
}
\source{
Statistics Finland \url{http://pxnet2.stat.fi/PXWeb/pxweb/en/StatFin/}.
}
\description{
Dataset containing number of deaths by drowning in Finland in 1969-2014,
yearly average summer temperatures (June to August) and
corresponding population sizes (in hundreds of thousands).
}
\keyword{datasets}
