% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfilter.R
\name{ekf_smoother}
\alias{ekf_smoother}
\title{Extended Kalman Smoothing}
\usage{
ekf_smoother(object)
}
\arguments{
\item{object}{Model object}
}
\value{
List containing the log-likelihood,
smoothed state estimates \code{alphahat}, and the corresponding variances \code{Vt} and
 \code{Ptt}.
}
\description{
Function \code{ekf_smoother} runs the extended Kalman smoother for the given 
non-linear Gaussian model of class \code{nlg_ssm}, 
and returns the filtered estimates and one-step-ahead predictions of the 
states \eqn{\alpha_t} given the data up to time \eqn{t}.
}

