% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{ekf_smoother}
\alias{ekf_smoother}
\title{Extended Kalman Smoothing}
\usage{
ekf_smoother(object, iekf_iter = 0)
}
\arguments{
\item{object}{Model object}

\item{iekf_iter}{If \code{iekf_iter > 0}, iterated extended Kalman filter is 
used with \code{iekf_iter} iterations.}
}
\value{
List containing the log-likelihood,
smoothed state estimates \code{alphahat}, and the corresponding variances \code{Vt} and
 \code{Ptt}.
}
\description{
Function \code{ekf_smoother} runs the (iterated) extended Kalman smoother for 
the given non-linear Gaussian model of class \code{nlg_ssm}, 
and returns the smoothed estimates of the states and the corresponding variances.
}
