% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssm-package.R
\docType{data}
\name{exchange}
\alias{exchange}
\title{Pound/Dollar daily exchange rates}
\format{
A vector of length 945.
}
\source{
The data used to be available on the www.ssfpack.com/DKbook.html but
this page is does not seem to be available anymore.
}
\description{
Dataset containing daily log-returns from 1/10/81-28/6/85 as in Durbin and
Koopman (2012).
}
\examples{
data("exchange")
model <- svm(exchange, rho = uniform(0.97,-0.999,0.999),
 sd_ar = halfnormal(0.175, 2), mu = normal(-0.87, 0, 2))

out <- particle_smoother(model, particles = 500)
plot.ts(cbind(model$y, exp(out$alphahat)))
}
\references{
James Durbin, Siem Jan Koopman (2012).
Time Series Analysis by State Space Methods. Oxford University Press.
https://doi.org/10.1093/acprof:oso/9780199641178.001.0001
}
\keyword{datasets}
