% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bssm-package.R
\docType{data}
\name{exchange}
\alias{exchange}
\title{Pound/Dollar daily exchange rates}
\format{
A vector of length 945.
}
\source{
\url{http://www.ssfpack.com/DKbook.html}.
}
\description{
Dataset containing daily log-returns from 1/10/81-28/6/85 as in [1]
}
\references{
James Durbin, Siem Jan Koopman (2012). "Time Series Analysis by State Space Methods". 
Oxford University Press.
}
\keyword{datasets}
