% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx.R
\name{gaussian_approx}
\alias{gaussian_approx}
\title{Gaussian Approximation of Non-Gaussian/Non-linear State Space Model}
\usage{
gaussian_approx(model, max_iter, conv_tol, ...)
}
\arguments{
\item{model}{Model to be approximated.}

\item{max_iter}{Maximum number of iterations.}

\item{conv_tol}{Tolerance parameter.}

\item{...}{Ignored.}

\item{iekf_iter}{For non-linear models, number of iterations in iterated EKF (defaults to 0).}
}
\description{
Returns the approximating Gaussian model.
}
