% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ng_bsm.R, R/ngssm.R, R/svm.R
\name{importance_sample}
\alias{importance_sample}
\alias{importance_sample.ng_bsm}
\alias{importance_sample.ngssm}
\alias{importance_sample.svm}
\title{Importance Sampling from non-Gaussian State Space Model}
\usage{
importance_sample(object, nsim, seed, ...)

\method{importance_sample}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{importance_sample}{ngssm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{importance_sample}{svm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{of class \code{ng_bsm}.}

\item{nsim}{Number of samples.}

\item{seed}{Seed for the random nunber generator.}

\item{...}{Ignored.}
}
\description{
Returns \code{nsim} samples from the approximating Gaussian model with corresponding (scaled) importance weights.
}

