% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R, R/gssm.R
\name{kfilter.bstsm}
\alias{kfilter}
\alias{kfilter.bstsm}
\alias{kfilter.gssm}
\title{Kalman Filtering}
\usage{
\method{kfilter}{bstsm}(object, ...)

kfilter(object, ...)

\method{kfilter}{gssm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gssm} or \code{bstsm}}

\item{...}{Ignored.}
}
\value{
List containing theone-step-ahead predictions \code{at} and filtered
estimates \code{att} of states, and the corresponding variances \code{Pt} and
 \code{Ptt}.
}
\description{
Kalman filtering of linear Gaussian models.
}
\details{
\code{kfilter} runs the Kalman filter for the given model
(and it's parameters), and returns the filtered estimates  and
one-step-ahead predictions of the states \eqn{\alpha_t} given the
model parameters and data up to time \eqn{t}.
}

