% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kfilter.R
\name{kfilter}
\alias{kfilter}
\title{Kalman Filtering}
\usage{
kfilter(model, ...)
}
\arguments{
\item{model}{Model Model object.}

\item{...}{Ignored.}
}
\value{
List containing the log-likelihood (approximate in non-Gaussian case),
one-step-ahead predictions \code{at} and filtered
estimates \code{att} of states, and the corresponding variances \code{Pt} and
 \code{Ptt}.
}
\description{
Function \code{kfilter} runs the Kalman filter for the given model, 
and returns the filtered estimates and one-step-ahead predictions of the 
states \eqn{\alpha_t} given the data up to time \eqn{t}.
}
\details{
For non-Gaussian models, the Kalman filtering is based on the approximate Gaussian model.
}
\seealso{
\code{\link{bootstrap_filter}}
}
