% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLik.gssm}
\alias{logLik.gssm}
\alias{logLik.ngssm}
\title{Log-likelihood of the State Space Model}
\usage{
\method{logLik}{gssm}(object, ...)

\method{logLik}{ngssm}(object, nsim_states, seed = 1, ...)
}
\arguments{
\item{object}{Model object.}

\item{...}{Ignored.}

\item{nsim_states}{Number of samples for importance sampling. If 0, approximate log-likelihood is returned.
See vignette for details.}

\item{seed}{Seed for the random number generator. Compared to other functions of the package, the
default seed is fixed (as 1) in order to work properly in numerical optimization algorithms.}
}
\description{
Computes the log-likelihood of the state space model of \code{bssm} package.
}

