% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglik.R
\name{logLik.gaussian}
\alias{logLik.gaussian}
\title{Log-likelihood of a Gaussian State Space Model}
\usage{
\method{logLik}{gaussian}(object, ...)
}
\arguments{
\item{object}{Model model.}

\item{...}{Ignored.}
}
\description{
Computes the log-likelihood of a linear-Gaussian state space model of \code{bssm} package.
}
\examples{
model <- ssm_ulg(y = c(1,4,3), Z = 1, H = 1, T = 1, R = 1)
logLik(model)
}
