% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/gssm.R, R/ng_bsm.R, R/ngssm.R
\name{logLik.bsm}
\alias{logLik.bsm}
\alias{logLik.gssm}
\alias{logLik.ng_bsm}
\alias{logLik.ngssm}
\title{Log-likelihood of Exponential Family State Space Model}
\usage{
\method{logLik}{bsm}(object, ...)

\method{logLik}{gssm}(object, ...)

\method{logLik}{ng_bsm}(object, nsim_states, seed = 1, ...)

\method{logLik}{ngssm}(object, nsim_states, seed = 1, ...)
}
\arguments{
\item{object}{Model object.}

\item{...}{Ignored.}

\item{nsim_states}{Number of samples for importance sampling. If 0, approximate log-likelihood is returned.}

\item{seed}{Seed for Boost random number generator. Compared to other functions of the package, here the
default seed is fixed (as 1) in order to work properly in optimization algorithms.}
}
\description{
Computes the log-likelihood of exponential family state space model.
}

