% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_bsm.R
\name{ng_bsm}
\alias{ng_bsm}
\title{Non-Gaussian Basic Structural (Time Series) Model}
\usage{
ng_bsm(y, sd_level, sd_slope, sd_seasonal, sd_noise, distribution, phi = 1,
  xreg = NULL, beta = NULL, period = frequency(y), slope = TRUE,
  seasonal = frequency(y) > 1, a1, P1, lower_prior, upper_prior)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{sd_level}{Standard error of level equation. Used as an initial
value in MCMC. If missing, \code{sd_level} is fixed to zero.}

\item{sd_slope}{Standard error of slope equation. Used as an initial
value in MCMC. If missing, \code{sd_slope} is fixed to zero.
Ignored if \code{slope = FALSE}.}

\item{sd_seasonal}{Standard error of seasonal equation. Used as an initial
value in MCMC. If missing, \code{sd_seasonal} is fixed to zero.}

\item{sd_noise}{Standard error of additional noise term. Used as an initial
value in MCMC. If missing, additional noise term is omitted from the model.}

\item{distribution}{distribution of the observation. Possible choices are
\code{"poisson"} and \code{"binomial"}.}

\item{phi}{Additional parameter vector relating to the non-Gaussian distribution.
For Poisson distribution, this corresponds to offset term. For binomial, this
is the number of trials.}

\item{xreg}{Matrix containing covariates.}

\item{beta}{Regression coefficients. Used as an initial
value in MCMC. Defaults to vector of zeros.}

\item{period}{Length of the seasonal component i.e. the number of
observations per season. Default is \code{frequency(y)}.}

\item{slope}{Should the model contain the slope term. Default is \code{TRUE}.
If \code{FALSE}, local level component is used instead of local trend component.}

\item{seasonal}{Should the model contain seasonal term. Default is
\code{frequency(y) > 1}.}

\item{a1}{Prior means for the initial states (level, slope, seasonals).
Defaults to vector of zeros.}

\item{P1}{Prior covariance for the initial states (level, slope, seasonals).
Default is diagonal matrix with 1e5 on the diagonal.}

\item{lower_prior, upper_prior}{Lower and upper bounds for the uniform prior
on standard deviations (sd_level, sd_slope, sd_seasonal) and regression
coefficients. Defaults to zero for lower bound and and
\code{sd(init_signal)} for upper bound of standard deviations and
(-1e4, 1e4) for regression coefficients.}
}
\value{
Object of class \code{ng_bsm}.
}
\description{
Constructs a non-Gaussian basic structural model with local level or
local trend component, a seasonal component, and regression component
(or subset of these components).
}
\examples{
model <- ng_bsm(Seatbelts[, "VanKilled"], distribution = "poisson",
  sd_level = 0.01, sd_seasonal = 0.01, slope = FALSE,
  xreg = Seatbelts[, "law"])
\dontrun{
set.seed(123)
mcmc_out <- run_mcmc(model, n_iter = 5000, nsim = 20)
mcmc_out$acceptance_rate
plot(mcmc_out$theta)
summary(mcmc_out$theta)

require("ggplot2")
ggplot(as.data.frame(mcmc_out$theta[,1:2]), aes(x = sd_level, y = sd_seasonal)) +
  geom_point() + stat_density2d(aes(fill = ..level.., alpha = ..level..),
  geom = "polygon") + scale_fill_continuous(low = "green",high = "blue") +
  guides(alpha = "none")

pred <- predict(model, n_iter = 5000, nsim_states = 25, n_ahead = 36,
  probs = seq(0.05, 0.95, by = 0.05), newdata = matrix(1, 36, 1),
  newphi = rep(1, 36))
autoplot(pred)
}
}

