% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{nlg_ssm}
\alias{nlg_ssm}
\title{General multivariate nonlinear Gaussian state space models}
\usage{
nlg_ssm(y, Z, H, T, R, Z_gn, T_gn, a1, P1, theta, known_params = NA,
  known_tv_params = matrix(NA), n_states, n_etas, log_prior_pdf,
  time_varying = rep(TRUE, 4), state_varying = rep(TRUE, 2),
  state_names = paste0("state", 1:n_states))
}
\arguments{
\item{y}{Observations as multivariate time series (or matrix) of length \eqn{n}.}

\item{Z_gn, T_gn}{An external pointers for the C++ functions which 
define the gradients of the corresponding model functions.}

\item{a1}{Prior mean for the initial state as a vector of length m.}

\item{P1}{Prior covariance matrix for the initial state as m x m matrix.}

\item{theta}{Parameter vector passed to all model functions.}

\item{state_names}{Names for the states.}

\item{Z_fn, H_fn, T_fn, R_fn}{An external pointers for the C++ functions which 
define the corresponding model functions.}

\item{prior_pdf}{An external pointer for the C++ function which 
computes the log-prior density given theta.}
}
\value{
Object of class \code{gssm}.
}
\description{
Constructs an object of class \code{nlg_ssm} by defining the corresponding terms
of the observation and state equation:
}
\details{
\deqn{y_t = Z(\alpha_t) + H(\alpha_t) \epsilon_t, (\textrm{observation equation})}
\deqn{\alpha_{t+1} = T(\alpha_t) + R(\alpha_t)\eta_t, (\textrm{transition equation})}

where \eqn{\epsilon_t \sim N(0, I_p)}, \eqn{\eta_t \sim N(0, I_m)} and
\eqn{\alpha_1 \sim N(a_1, P_1)} independently of each other, and functions
\eqn{Z, H, T, R} can depend on \eqn{\alpha_t} and parameter vector \eqn{\theta}.

Compared to other models, these general models need a bit more effort from 
the user, as you must provide the several small C++ snippets which define the 
model structure. See examples in ZZZ.
}

