% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R, R/ng_bsm.R, R/svm.R
\name{particle_filter.bsm}
\alias{particle_filter}
\alias{particle_filter.bsm}
\alias{particle_filter.ng_bsm}
\alias{particle_filter.svm}
\title{Particle Filtering}
\usage{
\method{particle_filter}{bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

particle_filter(object, nsim, seed, ...)

\method{particle_filter}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_filter}{svm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{of class \code{bsm}, \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{seed}{Seed for RNG.}

\item{...}{Ignored.}
}
\value{
A list containing samples, weights from the last time point, and an
estimate of log-likelihood.
}
\description{
Function \code{particle_filter} performs a bootstrap filtering with stratification
resampling.
}

