% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backwards_sampling.R
\name{particle_simulate}
\alias{particle_simulate}
\alias{particle_simulate.bsm}
\alias{particle_simulate.gssm}
\alias{particle_simulate.ng_bsm}
\alias{particle_simulate.ngssm}
\alias{particle_simulate.svm}
\title{Backwards sampling}
\usage{
particle_simulate(object, nsim, nsim_store, seed, ...)

\method{particle_simulate}{gssm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_simulate}{bsm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_simulate}{ngssm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_simulate}{ng_bsm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_simulate}{svm}(object, nsim, nsim_store = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{Model.}

\item{nsim}{Number of samples used in particle filtering.}

\item{nsim_store}{Number of samples to store.}

\item{seed}{Seed for RNG.}

\item{...}{Ignored.}
}
\description{
Function \code{particle_simulate} performs backwards sampling from
the conditional distribution of the states $p(alpha|y, theta)$.
}

