% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R, R/gssm.R, R/ngssm.R, R/svm.R
\name{particle_smoother.bsm}
\alias{particle_smoother}
\alias{particle_smoother.bsm}
\alias{particle_smoother.gssm}
\alias{particle_smoother.ngssm}
\alias{particle_smoother.svm}
\title{Particle Smoothing}
\usage{
\method{particle_smoother}{bsm}(object, nsim, method = "fs",
  seed = sample(.Machine$integer.max, size = 1), ...)

particle_smoother(object, nsim, method, seed, ...)

\method{particle_smoother}{gssm}(object, nsim, method = "fs",
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_smoother}{ngssm}(object, nsim, method = "fs",
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{particle_smoother}{svm}(object, nsim, method = "fs",
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{of class \code{bsm}, \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{method}{Either \code{"fs"} (filter-smoother), or \code{"fbs"} (forward-backward smoother).}

\item{seed}{Seed for RNG.}

\item{...}{Ignored.}
}
\description{
Function \code{particle_smoother} performs filter-smoother or forward-backward smoother,
using a bootstrap filtering with stratification resampling.
}

