% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R
\name{particle_smoother.bsm}
\alias{particle_smoother}
\alias{particle_smoother.bsm}
\title{Particle Smoothing}
\usage{
\method{particle_smoother}{bsm}(object, nsim, method = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

particle_smoother(object, nsim, seed, ...)
}
\arguments{
\item{object}{of class \code{bsm}, \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{seed}{Seed for RNG}

\item{...}{Ignored.}
}
\description{
Function \code{particle_smoother} performs a bootstrap filtering with stratification resampling, 
returning a weighted sample from $p(alpha|y, theta)$ and the estimate of log-likelihood.
}

