% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/ng_bsm.R, R/svm.R
\name{bootstrap_smoother}
\alias{bootstrap_smoother}
\alias{bootstrap_smoother.ng_bsm}
\alias{bootstrap_smoother.svm}
\alias{gap_smoother}
\alias{gap_smoother.ng_bsm}
\alias{gap_smoother.svm}
\alias{gap_smoother0}
\alias{gap_smoother0.ng_bsm}
\alias{gap_smoother0.svm}
\title{Particle Smoothing for non-Gaussian State Space Model}
\usage{
bootstrap_smoother(object, nsim, seed, ...)

gap_smoother(object, nsim, seed, ...)

gap_smoother0(object, nsim, seed, ...)

\method{bootstrap_smoother}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{gap_smoother}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{gap_smoother0}{ng_bsm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{bootstrap_smoother}{svm}(object, nsim,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{gap_smoother0}{svm}(object, nsim, seed = sample(.Machine$integer.max,
  size = 1), ...)

\method{gap_smoother}{svm}(object, nsim, seed = sample(.Machine$integer.max,
  size = 1), ...)
}
\arguments{
\item{object}{of class \code{ng_bsm} or \code{svm}.}

\item{nsim}{Number of samples.}

\item{seed}{Seed for RNG}

\item{...}{Ignored.}
}
\description{
Function \code{bootstrap_smoother} performs a bootstrap filtering with backward 
sampling, returning a weighted sample from $p(alpha|y, theta)$ and estimate of log-likelihood. 
Function \code{gap_smoother} first contructs a Gaussian approximation of the model, and uses the approximate
$p(alpha_1 | y)$ for simulating initial state in particle filtering.
}

