% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_bsm.R, R/ngssm.R
\name{predict.ng_bsm}
\alias{predict.ng_bsm}
\alias{predict.ngssm}
\title{Predictions for Non-Gaussian State Space Models}
\usage{
\method{predict}{ng_bsm}(object, n_iter, nsim_states, newdata = NULL,
  n_ahead = 1, interval = "mean", probs = c(0.05, 0.95),
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, seed = sample(.Machine$integer.max, size =
  1), newphi = NULL, ...)

\method{predict}{ngssm}(object, n_iter, nsim_states, lower_prior, upper_prior,
  newdata = NULL, n_ahead = 1, interval = "mean", probs = c(0.05, 0.95),
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, seed = sample(.Machine$integer.max, size =
  1), newphi = NULL, Z_est, T_est, R_est, ...)
}
\arguments{
\item{object}{Model object.}

\item{n_iter}{Number of MCMC iterations.}

\item{nsim_states}{Number of simulations of states per MCMC iteration. Only
used when \code{type = "full"}.}

\item{newdata}{Matrix containing the covariate values for the future time
points. Defaults to zero matrix of appropriate size.}

\item{n_ahead}{Number of steps ahead at which to predict.}

\item{interval}{Compute predictions on \code{"mean"} ("confidence interval") or
\code{"response"} ("prediction interval"). Defaults to \code{"response"}.}

\item{probs}{Desired quantiles. Defaults to \code{c(0.05, 0.95)}. Always includes median 0.5.}

\item{n_burnin}{Length of the burn-in period which is disregarded from the
results. Defaults to \code{n_iter / 2}.}

\item{n_thin}{Thinning rate. Defaults to 1. Increase for large models in
order to save memory.}

\item{gamma}{Tuning parameter for the adaptation of RAM algorithm. Must be
between 0 and 1 (not checked).}

\item{target_acceptance}{Target acceptance ratio for RAM. Defaults to 0.234.}

\item{S}{Initial value for the lower triangular matrix of RAM
algorithm, so that the covariance matrix of the Gaussian proposal
distribution is \eqn{SS'}.}

\item{seed}{Seed for the random number generator.}

\item{newphi}{Vector of length \code{n_ahead} defining the future values of \eqn{\phi}.
Defaults to 1, expect for negative binomial distribution, where the initial
value is taken from \code{object$phi}.}

\item{...}{Ignored.}

\item{lower_prior}{Bounds of the uniform prior for parameters
\eqn{\theta}. Optional for \code{bsm} objects.}

\item{upper_prior}{Bounds of the uniform prior for parameters
\eqn{\theta}. Optional for \code{bsm} objects.}

\item{Z_est, T_est, R_est}{Matrices or arrays with same dimensions as the
corresponding system matrices in \code{object}, where \code{NA} values
identify the unknown parameters for estimation.}
}
\value{
List containing the mean predictions, quantiles and Monte Carlo
standard errors.
}
\description{
Posterior intervals of future observations or their means
(success probabilities in binomial case) for non-Gaussian models. These are
computed using either the quantile method where the intervals are computed
as empirical quantiles the posterior sample, or parametric method by
Helske (2016).
Note that for non-Gaussian models, the parametric method produces only approximate
intervals for the mean even when importance sampling is used, by assuming
that \eqn{p(\alpha | y)} is Gaussian.
}
\examples{
data(poisson_series)
model <- ng_bsm(poisson_series, sd_level = halfnormal(0.1, 1), 
  sd_slope=halfnormal(0.01, 0.1), distribution = "poisson")
pred <- predict(model, n_iter = 1e4, nsim = 10, n_ahead = 10, 
  probs = seq(0.05,0.95, by = 0.05)) 
autoplot(pred, median_color = "blue", mean_color = "red")

}

