% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/gssm.R
\name{predict.bstsm}
\alias{predict.bstsm}
\alias{predict.gssm}
\title{Predictions for Gaussian State Space Models}
\usage{
\method{predict}{bstsm}(object, n_iter, lower_prior, upper_prior,
  newdata = NULL, n_ahead = 1, interval = "response", probs = c(0.05,
  0.95), method = "parametric", return_MCSE = TRUE, nsim_states = 1,
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, seed = sample(.Machine$integer.max, size =
  1), log_space = FALSE, ...)

\method{predict}{gssm}(object, n_iter, lower_prior, upper_prior,
  newdata = NULL, n_ahead = 1, interval = "response", probs = c(0.05,
  0.95), method = "parametric", return_MCSE = TRUE, nsim_states = 1,
  n_burnin = floor(n_iter/2), n_thin = 1, gamma = 2/3,
  target_acceptance = 0.234, S, Z_est, H_est, T_est, R_est,
  seed = sample(.Machine$integer.max, size = 1), ...)
}
\arguments{
\item{object}{Object of class \code{gssm} or \code{bstsm}.}

\item{n_iter}{Number of MCMC iterations.}

\item{lower_prior}{Bounds of the uniform prior for parameters
\eqn{\theta}. Optional for \code{bstsm} objects.}

\item{upper_prior}{Bounds of the uniform prior for parameters
\eqn{\theta}. Optional for \code{bstsm} objects.}

\item{newdata}{Matrix containing the covariate values for the future time
points.}

\item{n_ahead}{Number of steps ahead at which to predict.}

\item{interval}{Compute predictions on \code{"mean"} ("confidence interval") or
\code{"response"} ("prediction interval"). Defaults to \code{"response"}.}

\item{probs}{Desired quantiles. Defaults to \code{c(0.05, 0.95)}.}

\item{method}{Either \code{"parametric"} (default) or \code{"quantile"}.}

\item{nsim_states}{Number of simulations of states per MCMC iteration. Only
used when \code{type = "full"}.}

\item{n_burnin}{Length of the burn-in period which is disregarded from the
results. Defaults to \code{n_iter / 10}.}

\item{n_thin}{Thinning rate. Defaults to 1. Increase for long time series in
order to save memory.}

\item{gamma}{Tuning parameter for the adaptation of RAM algorithm. Must be
between 0 and 1 (not checked).}

\item{target_acceptance}{Target acceptance ratio for RAM.}

\item{S}{Initial value for the lower triangular matrix of RAM
algorithm, so that the covariance matrix of the Gaussian proposal
distribution is \eqn{SS'}.}

\item{...}{Ignored.}

\item{Z_est}{Arrays or matrices containing \code{NA}
values marking the unknown parameters which are to be estimated. Must be of
same dimension as the corresponding elements of the model object.}

\item{H_est}{Arrays or matrices containing \code{NA}
values marking the unknown parameters which are to be estimated. Must be of
same dimension as the corresponding elements of the model object.}

\item{T_est}{Arrays or matrices containing \code{NA}
values marking the unknown parameters which are to be estimated. Must be of
same dimension as the corresponding elements of the model object.}

\item{R_est}{Arrays or matrices containing \code{NA}
values marking the unknown parameters which are to be estimated. Must be of
same dimension as the corresponding elements of the model object.}
}
\value{
List containing the mean predictions, quantiles and Monte Carlo
standard errors.
}
\description{
Posterior intervals of future observations or their means
for Gaussian models. These are
computed using either the quantile method where the intervals are computed
as empirical quantiles the posterior sample, or parametric method by
Helske (2016).
}

