% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcmc.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Bayesian Inference of State Space Models}
\usage{
run_mcmc(object, n_iter, ...)
}
\arguments{
\item{object}{State space model object of \code{bssm} package.}

\item{n_iter}{Number of MCMC iterations.}

\item{...}{Parameters to specific methods. See \code{\link{run_mcmc.gssm}} and 
\code{\link{run_mcmc.ngssm}} for details.}
}
\description{
Adaptive Markov chain Monte Carlo simulation of state space models using
Robust Adaptive Metropolis algorithm by Vihola (2012).
}
\references{
Matti Vihola (2012). "Robust adaptive Metropolis algorithm with
coerced acceptance rate". Statistics and Computing, Volume 22, Issue 5,
pages 997--1008.
Matti Vihola, Jouni Helske, Jordan Franks (2016). "Importance sampling type 
correction of Markov chain Monte Carlo and exact approximations."
ArXiv:1609.02541.
}

