% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Bayesian Inference of State Space Models using MCMC with RAM}
\usage{
run_mcmc(object, ...)
}
\arguments{
\item{object}{Object of class \code{gssm}, \code{bsm}, \code{ngssm}, \code{svm},
or \code{ng_bsm}.}

\item{...}{Arguments to be passed to methods.
See \code{\link{run_mcmc.gssm}} and \code{\link{run_mcmc.ngssm}} for details.}
}
\description{
Adaptive Markov chain Monte Carlo simulation of state space models using
Robust Adaptive Metropolis algorithm by Vihola (2012).
}
\references{
Vihola, Matti (2012). "Robust adaptive Metropolis algorithm with
coerced acceptance rate". Statistics and Computing, Volume 22, Issue 5,
pages 997--1008.
}

