% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ng_bsm.R, R/ngssm.R, R/svm.R
\name{run_mcmc.ng_bsm}
\alias{run_mcmc.ng_bsm}
\alias{run_mcmc.ngssm}
\alias{run_mcmc.svm}
\title{Bayesian Inference of State Space Models using MCMC with RAM}
\usage{
\method{run_mcmc}{ng_bsm}(object, n_iter, nsim_states = 1, type = "full",
  lower_prior, upper_prior, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S,
  method = "delayed acceptance", log_space = FALSE, n_threads = 1,
  seed = sample(.Machine$integer.max, size = 1),
  thread_seeds = sample(.Machine$integer.max, size = n_threads), ...)

\method{run_mcmc}{ngssm}(object, n_iter, Z_est, T_est, R_est, lower_prior,
  upper_prior, nsim_states = 1, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{run_mcmc}{svm}(object, n_iter, nsim_states = 1, type = "full",
  lower_prior, upper_prior, n_burnin = floor(n_iter/2), n_thin = 1,
  gamma = 2/3, target_acceptance = 0.234, S,
  method = "delayed acceptance", n_threads = 1,
  seed = sample(.Machine$integer.max, size = 1),
  thread_seeds = sample(.Machine$integer.max, size = n_threads), ...)
}
\arguments{
\item{object}{Model object.}

\item{n_iter}{Number of MCMC iterations.}

\item{nsim_states}{Number of simulations of states per MCMC iteration. Only
used when \code{type = "full"}.}

\item{lower_prior, upper_prior}{Bounds of the uniform prior for parameters
\eqn{\theta}. Optional for \code{bsm} objects.}

\item{n_burnin}{Length of the burn-in period which is disregarded from the
results. Defaults to \code{n_iter / 2}.}

\item{n_thin}{Thinning rate. Defaults to 1. Increase for large models in
order to save memory.}

\item{gamma}{Tuning parameter for the adaptation of RAM algorithm. Must be
between 0 and 1 (not checked).}

\item{target_acceptance}{Target acceptance ratio for RAM. Defaults to 0.234.}

\item{S}{Initial value for the lower triangular matrix of RAM
algorithm, so that the covariance matrix of the Gaussian proposal
distribution is \eqn{SS'}.}

\item{method}{Use \code{"standard"} MCMC or \code{"delayed acceptance"} approach.}

\item{log_space}{Generate proposals for standard deviations in log-space. Default is \code{FALSE}.}

\item{seed}{Seed for the random number generator.}

\item{...}{Ignored.}

\item{Z_est, T_est, R_est}{Matrices or arrays with same dimensions as the
corresponding system matrices in \code{object}, where \code{NA} values
identify the unknown parameters for estimation.}
}
\description{
Adaptive Markov chain Monte Carlo simulation of state space models using
Robust Adaptive Metropolis algorithm by Vihola (2012).
}

