% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_smoother.R
\name{sim_smoother}
\alias{sim_smoother}
\alias{sim_smoother.gaussian}
\alias{sim_smoother.nongaussian}
\title{Simulation Smoothing}
\usage{
sim_smoother(model, nsim, seed, use_antithetic = FALSE, ...)

\method{sim_smoother}{gaussian}(
  model,
  nsim = 1,
  seed = sample(.Machine$integer.max, size = 1),
  use_antithetic = FALSE,
  ...
)

\method{sim_smoother}{nongaussian}(
  model,
  nsim = 1,
  seed = sample(.Machine$integer.max, size = 1),
  use_antithetic = FALSE,
  ...
)
}
\arguments{
\item{model}{Model object.}

\item{nsim}{Number of independent samples.}

\item{seed}{Seed for the random number generator.}

\item{use_antithetic}{Use an antithetic variable for location. 
Default is \code{FALSE}. Currently not used for multivariate models.}

\item{...}{Ignored.}
}
\value{
An array containing the generated samples.
}
\description{
Function \code{sim_smoother} performs simulation smoothing i.e. simulates 
the states from the conditional distribution \eqn{p(\alpha | y, \theta)} 
for linear-Gaussian models.
}
\details{
For non-Gaussian/non-linear models, the simulation is based on the 
approximating Gaussian model.
}
\examples{
# only missing data, simulates from prior
model <- bsm_lg(rep(NA, 25), sd_level = 1, 
  sd_y = 1)
# use antithetic variable for location
sim <- sim_smoother(model, nsim = 4, use_antithetic = TRUE, seed = 1)
ts.plot(sim[, 1, ])
cor(sim[, 1, ])
}
