% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_smoother.R
\name{sim_smoother}
\alias{sim_smoother}
\title{Simulation Smoothing}
\usage{
sim_smoother(object, nsim, seed, ...)
}
\arguments{
\item{object}{Model object.}

\item{nsim}{Number of independent samples.}

\item{seed}{Seed for the random number generator.}

\item{...}{Ignored.}
}
\value{
An array containing the generated samples.
}
\description{
Function \code{sim_smoother} performs simulation smoothing i.e. simulates the states
from the conditional distribution \eqn{p(\alpha | y, \theta)}.
}
\details{
For non-Gaussian models, the simulation smoothing is based on the approximate Gaussian model.
}
\examples{
model <- bsm(rep(NA, 50), sd_level = uniform(1,0,5), sd_y = uniform(1,0,5))
sim <- sim_smoother(model, 12)
ts.plot(sim[, 1, ])
}

