% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_smoother.R
\name{sim_smoother}
\alias{sim_smoother}
\title{Simulation Smoothing}
\usage{
sim_smoother(object, nsim, seed, use_antithetic = FALSE, ...)
}
\arguments{
\item{object}{Model object.}

\item{nsim}{Number of independent samples.}

\item{seed}{Seed for the random number generator.}

\item{use_antithetic}{Use an antithetic variable for location. Default is \code{FALSE}. Only used if \code{method} is "dk".}

\item{...}{Ignored.}

\item{method}{If \code{"dk"} (default), use simulation smoothing algorithm by Durbin and Koopman (2002). If \code{"psi"}, use twisted SMC. 
Only used for Gaussian models of class \code{"gssm"}, \code{"bsm"}, and \code{"ar1"}.}
}
\value{
An array containing the generated samples.
}
\description{
Function \code{sim_smoother} performs simulation smoothing i.e. simulates the states
from the conditional distribution \eqn{p(\alpha | y, \theta)}.
}
\details{
For non-Gaussian/non-linear models, the simulation is based on the approximating
Gaussian model.
}
\examples{
model <- bsm(rep(NA, 50), sd_level = uniform(1,0,5), sd_y = uniform(1,0,5))
sim <- sim_smoother(model, 12)
ts.plot(sim[, 1, ])
}
