% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{fast_smoother}
\alias{fast_smoother}
\alias{smoother}
\title{Kalman Smoothing}
\usage{
fast_smoother(model, ...)

smoother(model, ...)
}
\arguments{
\item{model}{Model model.}

\item{...}{Ignored.}
}
\value{
Matrix containing the smoothed estimates of states, or a list
with the smoothed states and the variances.
}
\description{
Methods for Kalman smoothing of the states. Function \code{fast_smoother}
computes only smoothed estimates of the states, and function
\code{smoother} computes also smoothed variances.
}
\details{
For non-Gaussian models, the smoothing is based on the approximate Gaussian model.
}
