% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsm.R, R/generics.R, R/gssm.R
\name{fast_smoother.bstsm}
\alias{fast_smoother}
\alias{fast_smoother.bstsm}
\alias{fast_smoother.gssm}
\alias{sim_smoother}
\alias{sim_smoother.bstsm}
\alias{sim_smoother.gssm}
\alias{smoother}
\alias{smoother.bstsm}
\alias{smoother.gssm}
\title{Kalman Smoothing}
\usage{
\method{fast_smoother}{bstsm}(object, ...)

\method{sim_smoother}{bstsm}(object, nsim = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{smoother}{bstsm}(object, ...)

fast_smoother(object, ...)

smoother(object, ...)

sim_smoother(object, nsim, ...)

\method{fast_smoother}{gssm}(object, ...)

\method{sim_smoother}{gssm}(object, nsim = 1,
  seed = sample(.Machine$integer.max, size = 1), ...)

\method{smoother}{gssm}(object, ...)
}
\arguments{
\item{object}{Object of class \code{gssm} or \code{bstsm}}

\item{...}{Ignored.}

\item{nsim}{Number of samples.}
}
\value{
Matrix containing the smoothed estimates of states, list
with the smoothed states and the variances, or an array  containing the
generated samples.
}
\description{
Methods for Kalman smoothing of the states. Function \code{fast_smoother}
computes only smoothed estimates of the states, and function
\code{smoother} computes also smoothed variances. Function
\code{sim_smoother} performs simulation smoothing i.e. simulates the states
from the conditional distribution \eqn{p(\alpha | y, \theta)}.
}

