% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{ssm_ung}
\alias{ssm_ung}
\title{General univariate non-Gaussian state space model}
\usage{
ssm_ung(
  y,
  Z,
  T,
  R,
  a1,
  P1,
  distribution,
  phi = 1,
  u = 1,
  init_theta = numeric(0),
  D,
  C,
  state_names,
  update_fn = default_update_fn,
  prior_fn = default_prior_fn
)
}
\arguments{
\item{y}{Observations as time series (or vector) of length \eqn{n}.}

\item{Z}{System matrix Z of the observation equation. Either a vector of length m,
a m x n matrix, or object which can be coerced to such.}

\item{T}{System matrix T of the state equation. Either a m x m matrix or a
m x m x n array, or object which can be coerced to such.}

\item{R}{Lower triangular matrix R the state equation. Either a m x k matrix or a
m x k x n array, or object which can be coerced to such.}

\item{a1}{Prior mean for the initial state as a vector of length m.}

\item{P1}{Prior covariance matrix for the initial state as m x m matrix.}

\item{distribution}{distribution of the observation. Possible choices are
\code{"poisson"}, \code{"binomial"}, and \code{"negative binomial"}.}

\item{phi}{Additional parameter relating to the non-Gaussian distribution.
For Negative binomial distribution this is the dispersion term, and for other
distributions this is ignored.}

\item{u}{Constant parameter for non-Gaussian models. For Poisson and negative binomial distribution, this corresponds to the offset
term. For binomial, this is the number of trials.}

\item{D}{Intercept terms \eqn{D_t} for the observations equation, given as a
1 x 1 or 1 x n matrix.}

\item{C}{Intercept terms \eqn{C_t} for the state equation, given as a
m times 1 or m times n matrix.}

\item{state_names}{Names for the states.}
}
\value{
Object of class \code{ssm_ung}.
}
\description{
Construct an object of class \code{ssm_ung} by defining the corresponding terms
of the observation and state equation:
}
\details{
\deqn{p(y_t | D_t + Z_t \alpha_t), (\textrm{observation equation})}
\deqn{\alpha_{t+1} = C_t + T_t \alpha_t + R_t \eta_t, (\textrm{transition equation})}

where \eqn{\eta_t \sim N(0, I_k)} and
\eqn{\alpha_1 \sim N(a_1, P_1)} independently of each other,
and \eqn{p(y_t | .)} is either Poisson, binomial or negative binomial distribution.
}
