% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmc.R
\name{summary.mcmc_output}
\alias{summary.mcmc_output}
\title{Summary of MCMC object}
\usage{
\method{summary}{mcmc_output}(x, only_theta = FALSE, ...)
}
\arguments{
\item{x}{output from \code{run_mcmc}}

\item{only_theta}{If \code{TRUE}, summaries are computed only for hyperparameters theta.}

\item{...}{Ignored.}
}
\description{
This functions returns a list containing mean, standard deviations, standard errors, and 
effective sample size estimates for parameters and states.
}
\details{
Note that computing the state summaries can be slow for large models due to repeated 
calls to \code{\link[coda]{spectrum0.ar}}.
}
