% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_mcmc.R
\name{summary.mcmc_output}
\alias{summary.mcmc_output}
\title{Summary of MCMC object}
\usage{
\method{summary}{mcmc_output}(object, return_se = FALSE, variable = "theta", only_theta = FALSE, ...)
}
\arguments{
\item{object}{Output from \code{run_mcmc}}

\item{return_se}{if \code{FALSE} (default), computation of standard 
errors and effective sample sizes is omitted. 
This saves time, as computing the spectral densities (by \code{coda}) can be slow for 
large models.}

\item{variable}{Are the summary statistics computed for either \code{"theta"} (default), 
\code{"states"}, or \code{"both"}?}

\item{only_theta}{Deprecated. If \code{TRUE}, summaries are computed only for hyperparameters theta.}

\item{...}{Ignored.}
}
\description{
This functions returns a list containing mean, standard deviations, standard errors, and 
effective sample size estimates for parameters and states.
}
\details{
Note that computing the state summaries can be slow for large models due to repeated 
calls to \code{\link[coda]{spectrum0.ar}}.
}
