% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{svm}
\alias{svm}
\title{Stochastic Volatility Model}
\usage{
svm(y, ar, sd_ar, sigma, xreg = NULL, beta = NULL, prior_sd = sd(y, na.rm
  = TRUE), lower_prior, upper_prior)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{sd_ar}{Standard error of AR-process. Used as an initial
value in MCMC.}

\item{sigma}{Sigma. Used as an initial value in MCMC.}

\item{prior_sd}{Standard deviation for the half-Normal prior for sigma. 
Default is \code{2*sd(y)}.}

\item{lower_prior, upper_prior}{Prior bounds for uniform priors on mean, phi, and sd_ar.}

\item{mean}{Mean. Used as an initial value in MCMC.}

\item{phi}{AR coefficient. Used as an initial
value in MCMC.}
}
\value{
Object of class \code{svm}.
}
\description{
Constructs a simple stochastic volatility model with Gaussian errors and first order
autoregressive signal.
}

