% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svm.R
\name{svm}
\alias{svm}
\title{Stochastic Volatility Model}
\usage{
svm(y, ar, sd_ar, sigma, beta, xreg = NULL)
}
\arguments{
\item{y}{Vector or a \code{\link{ts}} object of observations.}

\item{ar}{prior for autoregressive coefficient.}

\item{sd_ar}{Prior for the standard deviation of noise of the AR-process.}

\item{sigma}{Prior for sigma parameter of observation equation.}

\item{beta}{Prior for the regression coefficients.}

\item{xreg}{Matrix containing covariates.}
}
\value{
Object of class \code{svm}.
}
\description{
Constructs a simple stochastic volatility model with Gaussian errors and
first order autoregressive signal.
}

